@@ -1,14 +1,96 @@
+
+import 'package:jobr/data/models/contract_type.dart';
+import 'package:jobr/data/models/function_type.dart';
+import 'package:jobr/data/models/language.dart';
+import 'package:jobr/data/models/location_type.dart';
+
 import '../models/vacancy.dart';
 import '../services/api_service.dart';
 
-class VacanciesService {
-  final ApiService _apiService;
+class VacanciesService extends ApiService {
+  VacanciesService();
 
-  VacanciesService() : _apiService = ApiService();
+  Future<List<LocationType>> getLocationTypes() async {
+    try {
+      final response = await getApi(
+        'vacancies/locations',
+        cacheDuration: Duration(
+          minutes: 30,
+        ),
+      );
+      return (response.data as List)
+          .map((json) => LocationType.fromJson(json))
+          .toList();
+    } catch (e) {
+      throw Exception('Failed to fetch location: $e');
+    }
+  }
+
+  Future<List<ContractType>> getContractTypes() async {
+    try {
+      final response = await getApi(
+        'vacancies/contracts',
+        cacheDuration: Duration(
+          minutes: 30,
+        ),
+      );
+      return (response.data as List)
+          .map((json) => ContractType.fromJson(json))
+          .toList();
+    } catch (e) {
+      throw Exception('Failed to fetch contracts: $e');
+    }
+  }
+
+  Future<List<Language>> getLanguages() async {
+    try {
+      final response = await getApi(
+        'vacancies/languages',
+        cacheDuration: Duration(
+          minutes: 30,
+        ),
+      );
+      return (response.data as List)
+          .map((json) => Language.fromJson(json))
+          .toList();
+    } catch (e) {
+      throw Exception('Failed to fetch languages: $e');
+    }
+  }
+
+  Future<List<FunctionType>> getFunctionTypes() async {
+    try {
+      final response = await getApi(
+        'vacancies/functions',
+        cacheDuration: Duration(
+          minutes: 30,
+        ),
+      );
+      return (response.data as List)
+          .map((json) => FunctionType.fromJson(json))
+          .toList();
+    } catch (e) {
+      throw Exception('Failed to fetch functions: $e');
+    }
+  }
+
+  Future<void> createVacancy(Vacancy vacancy) async {
+    try {
+      await postApi(
+        'vacancies/vacancies',
+        data: vacancy.toJson(),
+      );
+    } catch (e) {
+      throw Exception('Failed to create vacancy: $e');
+    }
+  }
 
   Future<List<Vacancy>> getVacancies() async {
     try {
-      final response = await _apiService.dio.get('vacancies/');
+      final response = await getApi(
+        'vacancies/vacancies',
+      );
+
       return (response.data as List)
           .map((json) => Vacancy.fromJson(json))
           .toList();
@@ -19,7 +101,7 @@ class VacanciesService {
 
   Future<Vacancy> applyForVacancy(int vacancyId, int employeeId) async {
     try {
-      final response = await _apiService.dio.post(
+      final response = await postApi(
         'vacancies/$vacancyId/apply/',
         data: {
           'employee_id': employeeId,
