@@ -1,12 +1,13 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import 'package:jobr/core/routing/mixins/screen_state_mixin.dart';
 import 'package:jobr/data/models/user.dart';
 import 'package:jobr/features/authentication/screens/email_login_screen.dart';
 import 'package:jobr/features/authentication/screens/email_register_screen.dart';
+import 'package:jobr/features/authentication/screens/first_glance_screen.dart';
 import 'package:jobr/features/authentication/widgets/privacy_policy_block.dart';
-import 'package:jobr/ui/buttons/jobr_icon_button.dart';
+import 'package:jobr/ui/widgets/buttons/jobr_icon_button.dart';
 import 'package:jobr/ui/theme/jobr_icons.dart';
+import 'package:lyte_studios_flutter_ui/mixins/screen_state_mixin.dart';
 
 class LoginScreen extends StatefulWidget {
   final UserType userType;
@@ -19,7 +20,7 @@ class LoginScreen extends StatefulWidget {
     required this.isNewUser,
   });
 
-  static const String route = '/$location';
+  static const String route = '${FirstGlanceScreen.route}/$location';
   static const String location = 'login';
 
   @override
@@ -74,6 +75,9 @@ class _LoginScreenState extends State<LoginScreen> with ScreenStateMixin {
           },
         ),
         const PrivacyPolicyBlock(),
+        SizedBox(
+          height: 20,
+        )
       ],
     );
   }
