@@ -1,238 +1,202 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
 import 'package:go_router/go_router.dart';
+import 'package:jobr/core/routing/router.dart';
+import 'package:jobr/features/chat/screens/employee/chat_request_page_employee_screen.dart';
+import 'package:jobr/features/chat/screens/employer/chat_request_page_screen.dart';
+import 'package:jobr/ui/theme/padding_sizes.dart';
 import 'package:lyte_studios_flutter_ui/lyte_studios_flutter_ui.dart';
 
-import '../../../data/models/message.dart';
-
 class ChatRequestScreen extends StatefulWidget {
-  const ChatRequestScreen({super.key});
-
+  const ChatRequestScreen({super.key, this.isEmployeeSide = false});
+  final bool isEmployeeSide;
   static const String location = 'chat-request';
+
+  static String employerRoute = JobrRouter.getRoute(
+    location,
+    JobrRouter.employerInitialroute,
+  );
+
+  static String employeeRoute = JobrRouter.getRoute(
+    location,
+    JobrRouter.employeeInitialroute,
+  );
 
   @override
   State<ChatRequestScreen> createState() => _ChatScreenState();
 }
 
 class _ChatScreenState extends State<ChatRequestScreen> {
-  final List<Message> messages = [];
+  final List<Map<String, dynamic>> chatData = [
+    {
+      'name': 'Djimon Dilvoye',
+      'message': 'Hey! Ik zag op jullie website dat...',
+      'time': '07:23',
+      'image': 'assets/images/logos/image 15.png',
+      'unreadCount': 1,
+    },
+  ];
 
   @override
   Widget build(BuildContext context) {
-    final screenWidth = MediaQuery.of(context).size.width;
-    final screenHeight = MediaQuery.of(context).size.height;
     return Scaffold(
       appBar: AppBar(
-        automaticallyImplyLeading: false, // Removes the default back arrow
-        /* leadingWidth: 80,
-        leading: const Icon(Icons.arrow_back_ios,size: 30,color:  Color(0xFF000000)),*/
+        automaticallyImplyLeading: false,
         title: Row(
           children: [
-            const SizedBox(
-              width: 22,
-            ),
             ClearInkWell(
               onTap: () {
                 context.pop();
               },
-              child: const Icon(
-                Icons.arrow_back_ios,
-                size: 30,
-                color: Color(0xFF000000),
-              ),
-            ),
-            const SizedBox(
-              width: 10,
-            ),
-            Container(
-              width: 33,
-              height: 33,
-              decoration: const BoxDecoration(
-                shape: BoxShape.circle,
-                image: DecorationImage(
-                  image: AssetImage("assets/images/logos/spicy_lemon.png"),
-                  fit: BoxFit.cover,
+              child: const Padding(
+                padding: EdgeInsets.symmetric(
+                  horizontal: PaddingSizes.large,
+                ),
+                child: Icon(
+                  Icons.arrow_back_ios,
+                  size: 21,
+                  color: Color(0xFF000000),
                 ),
               ),
             ),
-            const SizedBox(
-              width: 10,
-            ),
-            const Text(
-              'Spicy Lemon',
-              style: TextStyle(
+            const SizedBox(width: 10),
+            Text(
+              'Berichtverzoeken',
+              style: const TextStyle(
                 fontSize: 20,
                 fontWeight: FontWeight.w600,
               ),
             ),
           ],
         ),
-        actions: [
-          Padding(
-            padding: const EdgeInsets.only(right: 13.0),
-            child: SvgPicture.asset(
-              "assets/images/logos/info.svg",
-              height: 28,
-              width: 28,
-            ),
-          )
-        ],
       ),
-      body: Container(
-        width: screenWidth,
-        height: screenHeight,
-        color: Colors.white,
+      body: SafeArea(
         child: Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 17.0),
+          padding: const EdgeInsets.symmetric(horizontal: 16.0),
           child: Column(
             children: [
+              Text(
+                widget.isEmployeeSide
+                    ? 'Hier vind je bedrijven die jouw profiel interessant vinden en willen chatten met jou.'
+                    : 'Hier vind je werkzoekenden die zelf contact opnemen. Dit kan een vraag zijn of een spontane sollicitatie.',
+                style: TextStyle(
+                    fontSize: 15,
+                    color: Colors.black54,
+                    fontWeight: FontWeight.w500),
+              ),
+              const SizedBox(height: 18),
               const Divider(
                 color: Color(0x0F000000),
                 height: 20,
-                thickness: 2,
-              ),
-              const Text(
-                "Zaterdag",
-                style: TextStyle(fontSize: 16, color: Color(0xFF696969)),
-              ),
-              Container(
-                decoration: BoxDecoration(
-                  color: const Color(0xFFF6F6F6),
-                  borderRadius: BorderRadius.circular(17),
-                ),
-                padding: const EdgeInsets.all(16),
-                // Internal padding
-                margin: const EdgeInsets.symmetric(vertical: 8, horizontal: 16),
-                // External spacing
-                child: Row(
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: [
-                    const Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Text(
-                          'Winkelmedewerker',
-                          style: TextStyle(
-                            fontSize: 17,
-                            fontWeight: FontWeight.w600,
-                          ),
-                        ),
-                        Text(
-                          'Student',
-                          style: TextStyle(fontSize: 15, color: Colors.grey),
-                        ),
-                      ],
-                    ),
-                    ElevatedButton(
-                      onPressed: () {},
-                      style: ElevatedButton.styleFrom(
-                        backgroundColor: const Color(0xFFF6F6F6),
-                        // Button background color
-                        shape: RoundedRectangleBorder(
-                          borderRadius: BorderRadius.circular(19.21),
-                          side: const BorderSide(
-                              color: Color(0xFFFF3E68),
-                              width: 1.92), // Border style
-                        ),
-                      ),
-                      child: const Text(
-                        'Bekijk vacature',
-                        style: TextStyle(
-                          color: Color(0xFFFF3E68),
-                          fontSize: 16,
-                          fontWeight: FontWeight.w600,
-                        ),
-                      ),
-                    ),
-                  ],
-                ),
+                thickness: 1.5,
+                indent: 74,
               ),
               Expanded(
+                  child: GestureDetector(
+                onTap: () {
+                  if (widget.isEmployeeSide) {
+                    context.push(ChatRequestEmployeePageScreen.employeeRoute);
+                  } else {
+                    context.push(ChatRequestPageScreen.employerRoute);
+                  }
+                },
                 child: ListView.builder(
-                  itemCount: messages.length,
+                  itemCount: chatData.length,
                   itemBuilder: (context, index) {
-                    final message = messages[index];
+                    final chat = chatData[index];
                     return Column(
-                      // crossAxisAlignment: message.isSentByMe
-                      //     ? CrossAxisAlignment.end
-                      //     : CrossAxisAlignment.start,
                       children: [
-                        Align(
-                          // alignment: message.isSentByMe
-                          //     ? Alignment.centerRight
-                          //     : Alignment.centerLeft,
-                          child: Container(
-                            margin: const EdgeInsets.symmetric(vertical: 5.0),
-                            padding: const EdgeInsets.all(10.0),
-                            decoration: BoxDecoration(
-                              // color: message.isSentByMe
-                              //     ? const Color(0xFF3976FF)
-                              //     : const Color(0xFFF6F6F6),
-                              borderRadius: BorderRadius.circular(17.79),
-                            ),
-                            child: Text(
-                              message.content,
-                              style: const TextStyle(
-                                // color: message.isSentByMe
-                                //     ? Colors.white
-                                //     : Colors.black,
-                                fontSize: 16.74,
-                              ),
-                            ),
-                          ),
+                        _buildChatItem(chat),
+                        const Divider(
+                          color: Color(0x0F000000),
+                          height: 20,
+                          thickness: 1.5,
+                          indent: 74,
                         ),
-                        // if (message.showSeen == true)
-                        //   const Padding(
-                        //     padding: EdgeInsets.only(top: 2.0, right: 8.0),
-                        //     child: Text(
-                        //       "Verzonden",
-                        //       style: TextStyle(
-                        //         color: Color(0xFF6B6B6B),
-                        //         fontSize: 13,
-                        //       ),
-                        //     ),
-                        //   ),
                       ],
                     );
                   },
                 ),
-              ),
-              Row(
-                mainAxisAlignment: MainAxisAlignment.center,
-                children: [
-                  // Wrap the TextField with a SizedBox to adjust width
-                  const SizedBox(
-                    width: 333,
-                    height: 39,
-                    // Adjust the width as needed
-                    child: TextField(
-                      style: TextStyle(
-                          fontSize: 19.5,
-                          fontWeight: FontWeight.normal,
-                          color: Color(0xFFBABABA)),
-                      decoration: InputDecoration(
-                        hintText: 'Hier typen...',
-                        border: OutlineInputBorder(),
-                        contentPadding: EdgeInsets.symmetric(
-                            horizontal: 12.0), // Adjust the height
-                      ),
-                    ),
-                  ),
-                  const SizedBox(width: 6),
-                  SvgPicture.asset(
-                    "assets/images/logos/send_message.svg",
-                    height: 32,
-                    width: 32,
-                  )
-                ],
-              ),
-              const SizedBox(
-                height: 20,
-              )
+              )),
             ],
           ),
         ),
       ),
+    );
+  }
+
+  Widget _buildChatItem(Map<String, dynamic> chat) {
+    return Row(
+      crossAxisAlignment: CrossAxisAlignment.center,
+      children: [
+        // Profile Picture
+        Container(
+          width: 67,
+          height: 67,
+          decoration: BoxDecoration(
+            shape: BoxShape.circle,
+            image: DecorationImage(
+              image: AssetImage(chat['image']),
+              fit: BoxFit.cover,
+            ),
+          ),
+        ),
+        const SizedBox(width: 16),
+        // Chat Details
+        Expanded(
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Text(
+                chat['name'],
+                style: const TextStyle(
+                  fontSize: 17,
+                  fontWeight: FontWeight.w600,
+                ),
+              ),
+              const SizedBox(height: 4),
+              Text(
+                chat['message'],
+                style: const TextStyle(
+                  color: Color(0xFF8F8F8F),
+                  fontSize: 16,
+                  fontWeight: FontWeight.w500,
+                ),
+                overflow: TextOverflow.ellipsis,
+              ),
+            ],
+          ),
+        ),
+        const SizedBox(width: 16),
+        // Time and Unread Count
+        Column(
+          crossAxisAlignment: CrossAxisAlignment.end,
+          children: [
+            Text(
+              chat['time'],
+              style: TextStyle(
+                color: chat['unreadCount'] > 0
+                    ? const Color(0xFFFF3E68)
+                    : const Color(0xFF8F8F8F),
+                fontSize: 17,
+                fontWeight: FontWeight.bold,
+              ),
+            ),
+            if (chat['unreadCount'] > 0)
+              CircleAvatar(
+                radius: 10,
+                backgroundColor: const Color(0xFFFF3E68),
+                child: Text(
+                  '${chat['unreadCount']}',
+                  style: const TextStyle(
+                    color: Colors.white,
+                    fontSize: 12,
+                    fontWeight: FontWeight.bold,
+                  ),
+                ),
+              ),
+          ],
+        ),
+      ],
     );
   }
 }
