@@ -1,34 +1,41 @@
 import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
+import 'package:jobr/core/routing/router.dart';
+import 'package:jobr/data/models/message.dart';
+import 'package:jobr/data/services/chat_service.dart';
+import 'package:jobr/features/chat/screens/employee/chat_page_employee_screen.dart';
+import 'package:jobr/features/chat/screens/employer/chat_page_screen.dart';
 import 'package:jobr/features/chat/screens/chat_request_screen.dart';
 import 'package:jobr/ui/theme/padding_sizes.dart';
+import 'package:lyte_studios_flutter_ui/mixins/screen_state_mixin.dart';
+import 'package:lyte_studios_flutter_ui/theme/extensions/hex_color.dart';
 
 class ChatScreen extends StatefulWidget {
   static const String location = 'chat';
-
-  const ChatScreen({super.key});
+  final bool isEmployeeSide;
+  static String employeeRoute = JobrRouter.getRoute(
+    location,
+    JobrRouter.employeeInitialroute,
+  );
+  const ChatScreen({super.key, this.isEmployeeSide = false});
 
   @override
   State<ChatScreen> createState() => _ChatScreenState();
 }
 
-class _ChatScreenState extends State<ChatScreen> {
-  // Sample chat data
-  final List<Map<String, dynamic>> chatData = [
-    {
-      'name': 'Spicy Lemon',
-      'message': 'Top, tot dan!',
-      'time': '09:46',
-      'image': 'assets/images/logos/spicy_lemon.png',
-      'unreadCount': 0,
-    },
-    {
-      'name': 'Brooklyn Kortrijk',
-      'message': 'Ik: Plannen we anders..',
-      'time': '07:23',
-      'image': 'assets/images/logos/brooklyn_kortrijk.png',
-      'unreadCount': 1,
-    },
-  ];
+class _ChatScreenState extends State<ChatScreen> with ScreenStateMixin {
+  String selectedCategory = 'Alle';
+
+  List<Message> chatData = [];
+
+  @override
+  Future<void> loadData() async {
+    chatData = await ChatService().getMessages();
+
+    setState(() {
+      chatData = chatData;
+    });
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -70,12 +77,21 @@ class _ChatScreenState extends State<ChatScreen> {
                           const Color(0xFFFF3E68).withOpacity(0.9),
                         ),
                         const SizedBox(width: 6),
-                        _buildCategoryButton('Ongelezen',
-                            const Color(0xFFF8F8F8), const Color(0xFFA0A0A0)),
+                        _buildCategoryButton(
+                          'Ongelezen',
+                          const Color(0xFFF8F8F8),
+                          const Color(0xFFA0A0A0),
+                        ),
                       ],
                     ),
                     TextButton(
-                      onPressed: () {},
+                      onPressed: () {
+                        if (widget.isEmployeeSide) {
+                          context.push(ChatRequestScreen.employeeRoute);
+                        } else {
+                          context.push(ChatRequestScreen.employerRoute);
+                        }
+                      },
                       child: const Text(
                         '3 verzoeken',
                         style: TextStyle(
@@ -97,12 +113,11 @@ class _ChatScreenState extends State<ChatScreen> {
                 Expanded(
                     child: GestureDetector(
                   onTap: () {
-                    Navigator.push(
-                      context,
-                      MaterialPageRoute(
-                        builder: (context) => const ChatRequestScreen(),
-                      ),
-                    );
+                    if (widget.isEmployeeSide) {
+                      context.push(ChatPageEmployeeScreen.employeeRoute);
+                    } else {
+                      context.push(ChatPageScreen.employerRoute);
+                    }
                   },
                   child: ListView.builder(
                     itemCount: chatData.length,
@@ -131,9 +146,9 @@ class _ChatScreenState extends State<ChatScreen> {
   }
 
   // Helper method to build chat items
-  Widget _buildChatItem(Map<String, dynamic> chat) {
+  Widget _buildChatItem(Message chat) {
     return Row(
-      crossAxisAlignment: CrossAxisAlignment.center,
+      crossAxisAlignment: CrossAxisAlignment.start,
       children: [
         // Profile Picture
         Container(
@@ -142,7 +157,7 @@ class _ChatScreenState extends State<ChatScreen> {
           decoration: BoxDecoration(
             shape: BoxShape.circle,
             image: DecorationImage(
-              image: AssetImage(chat['image']),
+              image: AssetImage(chat.sender),
               fit: BoxFit.cover,
             ),
           ),
@@ -154,7 +169,8 @@ class _ChatScreenState extends State<ChatScreen> {
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
               Text(
-                chat['name'],
+                // chat['name'],
+                '',
                 style: const TextStyle(
                   fontSize: 17,
                   fontWeight: FontWeight.w600,
@@ -162,11 +178,13 @@ class _ChatScreenState extends State<ChatScreen> {
               ),
               const SizedBox(height: 4),
               Text(
-                chat['message'],
-                style: const TextStyle(
-                  color: Color(0xFF8F8F8F),
+                // chat['message'],
+                '',
+                style: TextStyle(
+                  color: HexColor.fromHex('8F8F8F'),
                   fontSize: 16,
                   fontWeight: FontWeight.w500,
+                  letterSpacing: 0,
                 ),
                 overflow: TextOverflow.ellipsis,
               ),
@@ -179,28 +197,29 @@ class _ChatScreenState extends State<ChatScreen> {
           crossAxisAlignment: CrossAxisAlignment.end,
           children: [
             Text(
-              chat['time'],
+              // chat['time'],
+              '',
               style: TextStyle(
-                color: chat['unreadCount'] > 0
-                    ? const Color(0xFFFF3E68)
-                    : const Color(0xFF8F8F8F),
+                // color: chat['unreadCount'] > 0
+                //     ? const Color(0xFFFF3E68)
+                //     : const Color(0xFF8F8F8F),
                 fontSize: 17,
-                fontWeight: FontWeight.bold,
+                fontWeight: FontWeight.w500,
               ),
             ),
-            if (chat['unreadCount'] > 0)
-              CircleAvatar(
-                radius: 10,
-                backgroundColor: const Color(0xFFFF3E68),
-                child: Text(
-                  '${chat['unreadCount']}',
-                  style: const TextStyle(
-                    color: Colors.white,
-                    fontSize: 12,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-              ),
+            // if (chat['unreadCount'] > 0)
+            //   CircleAvatar(
+            //     radius: 10,
+            //     backgroundColor: const Color(0xFFFF3E68),
+            //     child: Text(
+            //       '${chat['unreadCount']}',
+            //       style: const TextStyle(
+            //         color: Colors.white,
+            //         fontSize: 12,
+            //         fontWeight: FontWeight.bold,
+            //       ),
+            //     ),
+            //   ),
           ],
         ),
       ],
@@ -208,15 +227,47 @@ class _ChatScreenState extends State<ChatScreen> {
   }
 
   // Helper method to build category buttons
-  Widget _buildCategoryButton(String text, Color bgColor, Color textColor) {
+  Widget _buildCategoryButton(
+      String text, Color defaultBgColor, Color defaultTextColor) {
+    final bool isSelected = selectedCategory == text;
+    final bool isOngelenzenSelected = selectedCategory == 'Ongelezen';
+
+    // Define colors based on selection state
+    Color bgColor;
+    Color textColor;
+
+    if (text == 'Alle') {
+      if (isOngelenzenSelected) {
+        bgColor = const Color(0xFFF8F8F8).withOpacity(0.4);
+        textColor = const Color(0xFFA0A0A0);
+      } else {
+        bgColor = const Color(0xFFFFEEF1).withOpacity(0.4);
+        textColor = const Color(0xFFFF3E68).withOpacity(0.9);
+      }
+    } else {
+      // For 'Ongelezen' button
+      if (isSelected) {
+        bgColor = const Color(0xFFFFEEF1).withOpacity(0.4);
+        textColor = const Color(0xFFFF3E68).withOpacity(0.9);
+      } else {
+        bgColor = const Color(0xFFF8F8F8).withOpacity(0.4);
+        textColor = const Color(0xFFA0A0A0);
+      }
+    }
+
     return SizedBox(
       height: 35,
       width: text == 'Alle' ? 52 : 104,
-      // Adjust width for 'Alle' and 'Ongelezen'
       child: TextButton(
-        onPressed: () {},
+        onPressed: () {
+          setState(() {
+            if (selectedCategory != text) {
+              selectedCategory = text;
+            }
+          });
+        },
         style: TextButton.styleFrom(
-          backgroundColor: bgColor.withOpacity(0.4),
+          backgroundColor: bgColor,
           shape: RoundedRectangleBorder(
             borderRadius: BorderRadius.circular(11.0),
           ),
