@@ -1,10 +1,20 @@
 import 'package:flutter/material.dart';
+import 'package:jobr/core/routing/router.dart';
+import 'package:jobr/features/jobs/filter.dart';
+import 'package:jobr/features/jobs/job_screen.dart';
 import 'package:jobr/features/jobs/widgets/job_card.dart';
 import 'package:jobr/features/jobs/widgets/return_arrow.dart';
+import 'package:lyte_studios_flutter_ui/theme/extensions/hex_color.dart';
 
 class JobListScreen extends StatelessWidget {
+  static final String location = 'job-list';
+
   const JobListScreen({super.key});
 
+  static String employeeRoute = JobrRouter.getRoute(
+    '${JobScreen.location}/${FilterScreenEmployee.location}/$location',
+    JobrRouter.employeeInitialroute,
+  );
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -16,18 +26,19 @@ class JobListScreen extends StatelessWidget {
         title: const Text(
           "Zoekresultaten",
           style: TextStyle(
-              color: Colors.black, fontWeight: FontWeight.bold, fontSize: 16),
+            color: Colors.black,
+            fontWeight: FontWeight.bold,
+            fontSize: 20,
+          ),
         ),
         actions: [
           Padding(
             padding: const EdgeInsets.only(right: 16.0),
-            child: Container(
-              child: Image.asset(
-                height: 20,
-                width: 20,
-                color: Colors.black,
-                "assets/images/recruteren/filter.png",
-              ),
+            child: Image.asset(
+              height: 24,
+              width: 24,
+              color: Colors.black,
+              "assets/images/recruteren/filter.png",
             ),
           ),
         ],
@@ -65,41 +76,57 @@ class JobListScreen extends StatelessWidget {
   }
 }
 
-class FilterButton extends StatelessWidget {
+class FilterButton extends StatefulWidget {
   final String label;
 
   const FilterButton({super.key, required this.label});
+
+  @override
+  _FilterButtonState createState() => _FilterButtonState();
+}
+
+class _FilterButtonState extends State<FilterButton> {
+  bool isExpanded = false;
 
   @override
   Widget build(BuildContext context) {
     return Padding(
       padding:
           const EdgeInsets.only(left: 4.0), // Reduce padding around the button
-      child: Container(
-        padding: const EdgeInsets.symmetric(
-            horizontal: 8, vertical: 1), // Reduce padding inside the button
-        decoration: BoxDecoration(
-          border:
-              Border.all(color: Colors.pinkAccent, width: 1.3), // Pink border
-          borderRadius: BorderRadius.circular(20),
-          color: Colors.white, // White background
-        ),
-        child: Row(
-          children: [
-            Text(
-              label,
-              style: const TextStyle(
-                  color: Colors.pinkAccent,
+      child: InkWell(
+        onTap: () {
+          setState(() {
+            isExpanded = !isExpanded;
+          });
+        },
+        child: Container(
+          padding: const EdgeInsets.symmetric(
+              horizontal: 8, vertical: 2), // Reduce padding inside the button
+          decoration: BoxDecoration(
+            border: Border.all(
+                color: HexColor.fromHex('#FF3E68'), width: 1.3), // Pink border
+            borderRadius: BorderRadius.circular(20),
+            color: Colors.white, // White background
+          ),
+          child: Row(
+            children: [
+              Text(
+                widget.label,
+                style: TextStyle(
+                  color: HexColor.fromHex('#FF3E68'),
                   fontWeight: FontWeight.w500,
-                  fontSize: 14),
-            ),
-            const SizedBox(width: 4),
-            const Icon(
-              Icons.keyboard_arrow_down,
-              color: Colors.pinkAccent,
-              size: 25, // Slightly smaller icon
-            ),
-          ],
+                  fontSize: 15.45,
+                ),
+              ),
+              Icon(
+                isExpanded
+                    ? Icons.keyboard_arrow_up
+                    : Icons.keyboard_arrow_down,
+                color: HexColor.fromHex('#FF3E68'),
+                size: 27, // Slightly smaller icon
+              ),
+            ],
+          ),
         ),
       ),
     );
