@@ -1,5 +1,9 @@
 import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
+import 'package:jobr/core/routing/router.dart';
+import 'package:jobr/features/jobs/filter.dart';
 import 'package:jobr/features/jobs/job_listing.dart';
+import 'package:jobr/features/jobs/job_screen.dart';
 import 'package:jobr/features/jobs/widgets/return_arrow.dart';
 
 class JobDetailScreen extends StatelessWidget {
@@ -14,6 +18,13 @@ class JobDetailScreen extends StatelessWidget {
     required this.image,
   });
 
+  static const String location = 'category';
+
+  static String employeeRoute = JobrRouter.getRoute(
+    '${JobScreen.location}/$location',
+    JobrRouter.employeeInitialroute,
+  );
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -21,12 +32,17 @@ class JobDetailScreen extends StatelessWidget {
         actions: [
           Padding(
             padding: const EdgeInsets.only(right: 16.0),
-            child: Container(
-              child: Image.asset(
-                height: 20,
-                width: 20,
-                color: Colors.black,
-                "assets/images/recruteren/filter.png",
+            child: GestureDetector(
+              onTap: () {
+                context.push(FilterScreenEmployee.employeeRoute);
+              },
+              child: Container(
+                child: Image.asset(
+                  height: 23,
+                  width: 23,
+                  color: Colors.black,
+                  "assets/images/recruteren/filter.png",
+                ),
               ),
             ),
           ),
@@ -40,7 +56,7 @@ class JobDetailScreen extends StatelessWidget {
             Text(
               title,
               style: const TextStyle(
-                fontSize: 18,
+                fontSize: 20,
                 fontWeight: FontWeight.w700,
               ),
             ),
@@ -56,6 +72,7 @@ class JobDetailScreen extends StatelessWidget {
               mainAxisAlignment: MainAxisAlignment.start,
               children: [
                 FilterButton(label: "Nieuwste eerst"),
+                SizedBox(width: 2),
                 FilterButton(label: "€ Laag-Hoog"),
               ],
             ),
