@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:jobr/core/routing/router.dart';
 import 'package:jobr/ui/theme/text_styles.dart';
 import 'package:lyte_studios_flutter_ui/lyte_studios_flutter_ui.dart';
@@ -16,7 +17,7 @@ class JobrNavigationItem {
   });
 }
 
-class BaseNavBarScreen extends StatelessWidget {
+class BaseNavBarScreen extends StatefulWidget {
   final Widget child;
 
   final int? selectedIndex;
@@ -31,37 +32,56 @@ class BaseNavBarScreen extends StatelessWidget {
   });
 
   @override
+  BaseNavBarScreenState createState() => BaseNavBarScreenState();
+}
+
+class BaseNavBarScreenState extends State<BaseNavBarScreen> {
+  bool _isNavBarVisible = true;
+
+  void toggleNavBarVisibility(bool isVisible) {
+    setState(() {
+      _isNavBarVisible = isVisible;
+    });
+  }
+
+  @override
   Widget build(BuildContext context) {
     return Scaffold(
-      bottomNavigationBar: selectedIndex == null
-          ? null
-          : BottomNavigationBar(
+      bottomNavigationBar: _isNavBarVisible && widget.selectedIndex != null
+          ? BottomNavigationBar(
+              unselectedItemColor: HexColor.fromHex('#999999'),
               type: BottomNavigationBarType.fixed,
-              currentIndex: selectedIndex ?? 0,
+              currentIndex: widget.selectedIndex ?? 0,
               onTap: (int index) {
-                String route = routes[index].route;
+                HapticFeedback.lightImpact();
+
+                String route = widget.routes[index].route;
 
                 // Use GoRouter for navigation
                 router.pushReplacement(route);
               },
               selectedLabelStyle: TextStyles.bodySmall.copyWith(
                 color: Theme.of(context).primaryColor,
-                fontSize: 12,
+                fontWeight: FontWeight.w500,
+                fontSize: 13,
               ),
               unselectedLabelStyle: TextStyles.bodySmall.copyWith(
-                fontSize: 12,
+                fontSize: 13,
+                fontWeight: FontWeight.w500,
               ),
-              items: routes
+              items: widget.routes
                   .map((route) => BottomNavigationBarItem(
                         label: route.name,
                         icon: _NavigationBarIcon(
                           icon: route.icon,
-                          selected: selectedIndex == routes.indexOf(route),
+                          selected: widget.selectedIndex ==
+                              widget.routes.indexOf(route),
                         ),
                       ))
                   .toList(),
-            ),
-      body: child,
+            )
+          : null,
+      body: widget.child,
     );
   }
 }
