@@ -1,10 +1,18 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import 'package:jobr/features/profile/screens/company_screen/settings.dart';
+import 'package:go_router/go_router.dart';
+import 'package:jobr/features/profile/screens/company/edit_company_profile_screen.dart';
+import 'package:jobr/features/profile/screens/company_screen/base_navbar.dart';
+import 'package:jobr/features/profile/screens/company_screen/settings_screen.dart';
 import 'package:jobr/features/profile/screens/tabs/general_item_widget.dart';
 import 'package:jobr/ui/theme/jobr_icons.dart';
 import 'package:jobr/ui/theme/text_styles.dart';
 import 'package:lyte_studios_flutter_ui/ui/icons/svg_icon.dart';
+import 'package:google_maps_flutter/google_maps_flutter.dart';
+import 'package:geocoding/geocoding.dart';
+import 'package:flutter/foundation.dart'; // add this import
+import 'package:flutter/gestures.dart'; // add this import
+import 'package:url_launcher/url_launcher.dart'; // add this import
 
 class CompanyProfileScreen extends StatefulWidget {
   static const String location = 'company-profile';
@@ -22,6 +30,37 @@ class _CompanyProfileScreenState extends State<CompanyProfileScreen> {
     GeneralItemsWidget(),
   ];
 
+  GoogleMapController? _mapController;
+  LatLng _currentLocation = const LatLng(51.9225, 4.47917); // Default location
+  Set<Marker> _markers = {};
+
+  @override
+  void initState() {
+    super.initState();
+    _initializeLocation();
+  }
+
+  Future<void> _initializeLocation() async {
+    try {
+      List<Location> locations = await locationFromAddress(
+          "Kortrijk"); // Replace with actual company location
+      if (locations.isNotEmpty) {
+        setState(() {
+          _currentLocation =
+              LatLng(locations.first.latitude, locations.first.longitude);
+          _markers = {
+            Marker(
+              markerId: const MarkerId('companyLocation'),
+              position: _currentLocation,
+            ),
+          };
+        });
+      }
+    } catch (e) {
+      debugPrint('Error getting location: $e');
+    }
+  }
+
   @override
   Widget build(BuildContext context) {
     final theme = Theme.of(context);
@@ -37,6 +76,7 @@ class _CompanyProfileScreenState extends State<CompanyProfileScreen> {
                   expandedHeight: 200,
                   floating: false,
                   pinned: true,
+                  centerTitle: true,
                   title: innerBoxIsScrolled
                       ? const Text(
                           "Brooklyn Kortrijk",
@@ -50,22 +90,22 @@ class _CompanyProfileScreenState extends State<CompanyProfileScreen> {
                   scrolledUnderElevation: 0,
                   elevation: 0,
                   leading: innerBoxIsScrolled
-                      ? Padding(
-                          padding: const EdgeInsets.all(8.0),
-                          child: Container(
-                            width: 70,
-                            height: 70,
-                            decoration: BoxDecoration(
-                              shape: BoxShape.circle,
-                              border: Border.all(color: Colors.white, width: 4),
-                              image: const DecorationImage(
-                                image: AssetImage(
-                                    'assets/images/images/image-b.png'),
-                                fit: BoxFit.cover,
+                      ? Transform.translate(
+                          offset: const Offset(16.0,
+                              0.0), // Shift widget left without negative margin
+                          child: Padding(
+                            padding: const EdgeInsets.all(6.0),
+                            child: ClipOval(
+                              child: SizedBox(
+                                width: 40, // Set the exact size
+                                height: 40,
+                                child: Image.asset(
+                                  'assets/images/logos/brooklyn_kortrijk.png',
+                                  fit: BoxFit.cover,
+                                ),
                               ),
                             ),
-                          ),
-                        )
+                          ))
                       : null,
                   backgroundColor: theme.colorScheme.surface,
                   clipBehavior: Clip.none,
@@ -78,37 +118,78 @@ class _CompanyProfileScreenState extends State<CompanyProfileScreen> {
                           top: 0,
                           left: 0,
                           right: 0,
-                          child: Image.asset(
-                            'assets/images/images/profile_image.png',
-                            fit: BoxFit.cover,
-                            height: 180,
-                          ),
-                        ),
-                        Positioned(
-                          bottom: 0,
-                          left: 10,
-                          child: Container(
-                            width: 120,
-                            height: 120,
-                            decoration: BoxDecoration(
-                              shape: BoxShape.circle,
-                              border: Border.all(
-                                color: Colors.white,
-                                width: 4,
-                              ),
-                              image: const DecorationImage(
-                                image: AssetImage(
-                                    'assets/images/images/image-b.png'),
+                          child: Hero(
+                            tag: 'profileBackground',
+                            child: GestureDetector(
+                              onTap: () {
+                                showGeneralDialog(
+                                  context: context,
+                                  barrierDismissible: true,
+                                  barrierLabel: 'Dismiss',
+                                  barrierColor: Colors.black54,
+                                  pageBuilder: (_, __, ___) => ZoomImageDialog(
+                                    tag: 'profileBackground',
+                                    imagePath:
+                                        'assets/images/images/profile_image.png',
+                                  ),
+                                );
+                              },
+                              child: Image.asset(
+                                'assets/images/images/profile_image.png',
                                 fit: BoxFit.cover,
+                                height: 180,
                               ),
                             ),
                           ),
                         ),
                         Positioned(
-                          bottom: 25,
-                          right: 50,
+                          bottom: 0,
+                          left: 10,
+                          child: Hero(
+                            tag: 'profileLogo',
+                            child: GestureDetector(
+                              onTap: () {
+                                showGeneralDialog(
+                                  context: context,
+                                  barrierDismissible: true,
+                                  barrierLabel: 'Dismiss',
+                                  barrierColor: Colors.black54,
+                                  pageBuilder: (_, __, ___) => ZoomImageDialog(
+                                    height: 200,
+                                    width: 200,
+                                    tag: 'profileLogo',
+                                    imagePath:
+                                        'assets/images/logos/brooklyn_kortrijk.png',
+                                  ),
+                                );
+                              },
+                              child: Container(
+                                width: 120,
+                                height: 120,
+                                decoration: BoxDecoration(
+                                  shape: BoxShape.circle,
+                                  border: Border.all(
+                                    color: Colors.white,
+                                    width: 4,
+                                  ),
+                                  image: const DecorationImage(
+                                    image: AssetImage(
+                                      'assets/images/logos/brooklyn_kortrijk.png',
+                                    ),
+                                    fit: BoxFit.cover,
+                                  ),
+                                ),
+                              ),
+                            ),
+                          ),
+                        ),
+                        Positioned(
+                          bottom: 40,
+                          right: 60,
                           child: ElevatedButton.icon(
-                            onPressed: () {},
+                            onPressed: () {
+                              context.push(EditCompanyProfileScreen.route);
+                            },
                             style: ElevatedButton.styleFrom(
                               backgroundColor: theme.primaryColor,
                             ),
@@ -133,27 +214,22 @@ class _CompanyProfileScreenState extends State<CompanyProfileScreen> {
                           ),
                         ),
                         Positioned(
-                          bottom: 30,
+                          bottom: 43,
                           right: 10,
                           child: InkWell(
                             onTap: () {
                               // Add your settings action here
-                              Navigator.push(
-                                context,
-                                MaterialPageRoute(
-                                    builder: (context) =>
-                                        const SettingsScreen()),
-                              );
+                              context.push(SettingsScreen.route);
                             },
                             child: Container(
-                              padding: const EdgeInsets.all(8),
-                              decoration: const BoxDecoration(
+                              padding: const EdgeInsets.all(10),
+                              decoration: BoxDecoration(
                                 shape: BoxShape.circle,
-                                color: Colors.white,
+                                color: Colors.grey.shade100,
                               ),
                               child: SvgIcon(
                                 JobrIcons.settings1,
-                                size: 18.68,
+                                size: 20.68,
                                 color: TextStyles.disabledText,
                               ),
                             ),
@@ -180,25 +256,34 @@ class _CompanyProfileScreenState extends State<CompanyProfileScreen> {
                 const SizedBox(height: 4),
                 Row(
                   children: [
-                    Text(
-                      "wwww.brooklyn.be",
-                      style: TextStyle(
-                        fontSize: 17.01,
-                        fontWeight: FontWeight.w500,
-                        fontFamily: 'Inter',
-                        color: TextStyles.brooklyn,
+                    GestureDetector(
+                      onTap: () {
+                        launchUrl(Uri.parse("http://www.brooklyn.be"));
+                      },
+                      child: Text(
+                        "www.brooklyn.be",
+                        style: TextStyle(
+                          fontSize: 17.01,
+                          fontWeight: FontWeight.w500,
+                          fontFamily: 'Poppins',
+                          color: TextStyles.brooklyn,
+                        ),
                       ),
                     ),
                     const SizedBox(width: 16),
-                    SvgIcon(
-                      JobrIcons.web,
-                      size: 11.72,
-                      leaveUnaltered: true,
-                      color: TextStyles.brooklyn,
+                    GestureDetector(
+                      onTap: () {
+                        launchUrl(Uri.parse("http://www.brooklyn.be"));
+                      },
+                      child: SvgIcon(
+                        JobrIcons.web,
+                        size: 12.72,
+                        leaveUnaltered: true,
+                        color: TextStyles.brooklyn,
+                      ),
                     ),
                   ],
                 ),
-                const SizedBox(height: 8),
                 Row(
                   children: [
                     SvgIcon(
@@ -236,29 +321,29 @@ class _CompanyProfileScreenState extends State<CompanyProfileScreen> {
                     ),
                   ],
                 ),
-                const SizedBox(height: 16),
+                const SizedBox(height: 12),
                 Row(
                   children: [
                     SvgIcon(
                       JobrIcons.instagram,
                       size: 27.86,
-                      color: TextStyles.selectedText,
+                      color: Colors.grey.shade600,
                     ),
                     const SizedBox(width: 16),
                     SvgIcon(
                       JobrIcons.tiktok,
                       size: 27.86,
-                      color: TextStyles.disabledText,
+                      color: Colors.grey.shade600,
                     ),
                     const SizedBox(width: 16),
                     SvgIcon(
                       JobrIcons.facebook,
                       size: 27.86,
-                      color: TextStyles.disabledText,
+                      color: Colors.grey.shade600,
                     ),
                   ],
                 ),
-                const SizedBox(height: 16),
+                const SizedBox(height: 12),
                 Text(
                   "Multibrandstore & Webshop. Brooklyn, da's een mix van merken en heel veel broeken. Dat laatste nemen we als broekspeciaalzaak wel heel serieus. Kom zeker eens langs om het zelf te zeggen.",
                   style: TextStyle(
@@ -310,15 +395,55 @@ class _CompanyProfileScreenState extends State<CompanyProfileScreen> {
                   ),
                 ),
                 const SizedBox(height: 16),
+                GestureDetector(
+                  onTap: () {
+                    toggleNavBarVisibility(false);
+                    showLocationSelector(context).then((_) {
+                      toggleNavBarVisibility(true);
+                    });
+                  },
+                  child: Container(
+                    padding: const EdgeInsets.symmetric(
+                        horizontal: 16, vertical: 16),
+                    decoration: BoxDecoration(
+                      color: TextStyles.darkGray,
+                      borderRadius: BorderRadius.circular(17),
+                    ),
+                    child: Row(
+                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                      children: [
+                        const Text(
+                          "Wisselen van vestiging",
+                          style: TextStyle(
+                              fontSize: 15.2,
+                              fontFamily: 'Inter',
+                              fontWeight: FontWeight.w600,
+                              color: Colors.black),
+                        ),
+                        Row(
+                          children: [
+                            SvgIcon(
+                              JobrIcons.venueLocation,
+                              size: 20,
+                              color: Theme.of(context).primaryColor,
+                            ),
+                            const SizedBox(width: 4),
+                          ],
+                        ),
+                      ],
+                    ),
+                  ),
+                ),
+                const SizedBox(height: 16),
                 SizedBox(
-                  height: 200,
+                  height: 280,
                   child: ListView(
                     scrollDirection: Axis.horizontal,
                     physics: const BouncingScrollPhysics(),
                     children: [
                       Container(
                         margin: const EdgeInsets.only(right: 16),
-                        width: 300,
+                        width: 290,
                         child: ClipRRect(
                           borderRadius: BorderRadius.circular(12),
                           child: Image.asset(
@@ -344,31 +469,245 @@ class _CompanyProfileScreenState extends State<CompanyProfileScreen> {
                     ],
                   ),
                 ),
-                const SizedBox(height: 24),
+                const SizedBox(height: 12),
                 Container(
-                  margin: const EdgeInsets.symmetric(horizontal: 2),
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(12),
-                    boxShadow: [
-                      BoxShadow(
-                        color: Colors.grey.withOpacity(0.3),
-                        spreadRadius: 2,
-                        blurRadius: 5,
-                        offset:
-                            const Offset(0, 3), // changes position of shadow
-                      ),
-                    ],
-                  ),
+                  height: 200,
+                  margin:
+                      const EdgeInsets.symmetric(vertical: 12.0, horizontal: 2),
                   child: ClipRRect(
                     borderRadius: BorderRadius.circular(12),
-                    child: Image.asset(
-                      'assets/images/images/Frame-m.png',
-                      // Replace with the actual map image path
-                      fit: BoxFit.cover,
+                    child: GoogleMap(
+                      onMapCreated: (controller) {
+                        _mapController = controller;
+                      },
+                      initialCameraPosition: CameraPosition(
+                        target: _currentLocation,
+                        zoom: 14.0,
+                      ),
+                      markers: _markers,
+                      zoomControlsEnabled: false,
+                      mapToolbarEnabled: false,
+                      myLocationButtonEnabled: false,
+                      gestureRecognizers: <Factory<
+                          OneSequenceGestureRecognizer>>{
+                        Factory<EagerGestureRecognizer>(
+                            () => EagerGestureRecognizer()),
+                      },
                     ),
                   ),
                 ),
               ],
             )));
+  }
+
+  void toggleNavBarVisibility(bool isVisible) {
+    final baseNavBarState =
+        context.findAncestorStateOfType<BaseNavBarScreenState>();
+    baseNavBarState?.toggleNavBarVisibility(isVisible);
+  }
+
+  Future<void> showLocationSelector(BuildContext context) async {
+    // Hide the navigation bar
+    toggleNavBarVisibility(false);
+
+    await showModalBottomSheet(
+      context: context,
+      shape: const RoundedRectangleBorder(
+        borderRadius: BorderRadius.vertical(
+          top: Radius.circular(20),
+        ),
+      ),
+      builder: (context) {
+        return DraggableScrollableSheet(
+          initialChildSize: 0.70,
+          minChildSize: 0.25,
+          maxChildSize: 0.75,
+          expand: false,
+          builder: (context, scrollController) {
+            return SingleChildScrollView(
+              controller: scrollController,
+              child: Container(
+                padding:
+                    const EdgeInsets.symmetric(vertical: 16, horizontal: 16),
+                child: Column(
+                  mainAxisSize: MainAxisSize.min,
+                  children: [
+                    // Drag handle at the top
+                    Container(
+                      width: 40,
+                      height: 4,
+                      margin: const EdgeInsets.only(bottom: 16),
+                      decoration: BoxDecoration(
+                        color: Colors.grey.shade400,
+                        borderRadius: BorderRadius.circular(2),
+                      ),
+                    ),
+                    // Location 1
+                    _buildLocationTile(
+                      context,
+                      title: "Brooklyn",
+                      subtitle: "Brugge, Leistraat",
+                      onPressed: () {
+                        Navigator.pop(context);
+                        // Handle selection logic for "Brooklyn Brugge"
+                      },
+                    ),
+                    const SizedBox(height: 12),
+                    // Location 2
+                    _buildLocationTile(
+                      context,
+                      title: "Brooklyn",
+                      subtitle: "Gent, Voorstraat",
+                      onPressed: () {
+                        Navigator.pop(context);
+                        // Handle selection logic for "Brooklyn Gent"
+                      },
+                    ),
+                    const SizedBox(height: 12),
+                    // Location 3
+                    _buildLocationTile(
+                      context,
+                      title: "Brooklyn",
+                      subtitle: "Mechelen, Bruul",
+                      onPressed: () {
+                        Navigator.pop(context);
+                        // Handle selection logic for "Brooklyn Mechelen"
+                      },
+                    ),
+                  ],
+                ),
+              ),
+            );
+          },
+        );
+      },
+    );
+
+    // Show the navigation bar again
+    toggleNavBarVisibility(true);
+  }
+
+  Widget _buildLocationTile(
+    BuildContext context, {
+    required String title,
+    required String subtitle,
+    required VoidCallback onPressed,
+  }) {
+    return Container(
+      padding: const EdgeInsets.symmetric(vertical: 12, horizontal: 16),
+      decoration: BoxDecoration(
+        color: Colors.grey.shade100,
+        borderRadius: BorderRadius.circular(12),
+      ),
+      child: Row(
+        children: [
+          // Leading icon
+          CircleAvatar(
+              backgroundColor: Colors.blue,
+              radius: 24,
+              child: ClipRect(
+                child: Container(
+                  width: 70,
+                  height: 70,
+                  decoration: BoxDecoration(
+                    shape: BoxShape.circle,
+                    border: Border.all(color: Colors.white, width: 4),
+                    image: const DecorationImage(
+                      image: AssetImage(
+                          'assets/images/logos/brooklyn_kortrijk.png'),
+                      fit: BoxFit.cover,
+                    ),
+                  ),
+                ),
+              )),
+          const SizedBox(width: 16),
+          // Title and subtitle
+          Expanded(
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: [
+                Text(
+                  title,
+                  style: const TextStyle(
+                    fontWeight: FontWeight.bold,
+                    fontSize: 16,
+                  ),
+                ),
+                const SizedBox(height: 4),
+                Row(
+                  children: [
+                    SvgIcon(
+                      JobrIcons.location,
+                      size: 16,
+                      leaveUnaltered: true,
+                      color: Theme.of(context).primaryColor,
+                    ),
+                    const SizedBox(width: 4),
+                    Text(
+                      subtitle,
+                      style: TextStyle(
+                        fontSize: 14,
+                        color: Colors.grey.shade700,
+                      ),
+                    ),
+                  ],
+                ),
+              ],
+            ),
+          ),
+          const SizedBox(width: 16),
+          // "Kies" button
+          TextButton(
+            onPressed: onPressed,
+            child: Text(
+              "Kies",
+              style: TextStyle(
+                fontSize: 17,
+                fontWeight: FontWeight.bold,
+                color: Theme.of(context).primaryColor,
+              ),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+class ZoomImageDialog extends StatelessWidget {
+  final String tag;
+  final String imagePath;
+  final double? height;
+
+  final double? width;
+  const ZoomImageDialog(
+      {super.key,
+      required this.tag,
+      required this.imagePath,
+      this.height,
+      this.width});
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTap: () => Navigator.pop(context),
+      child: Material(
+        color: Colors.black54,
+        child: Center(
+          child: Hero(
+            tag: tag,
+            child: ClipRRect(
+              borderRadius: BorderRadius.circular(12),
+              child: (height != null)
+                  ? Image.asset(imagePath,
+                      height: height ?? 200,
+                      width: width ?? 200,
+                      fit: BoxFit.cover)
+                  : Image.asset(imagePath, fit: BoxFit.cover),
+            ),
+          ),
+        ),
+      ),
+    );
   }
 }
