@@ -1,9 +1,14 @@
 import 'package:flutter/material.dart';
-import 'package:jobr/ui/widget/common_button.dart';
+import 'package:go_router/go_router.dart';
+import 'package:jobr/features/profile/screens/employee_profile_screen_display.dart';
+import 'package:jobr/ui/theme/text_styles.dart';
+import 'package:jobr/ui/widgets/buttons/action_button.dart';
+import 'package:lyte_studios_flutter_ui/lyte_studios_flutter_ui.dart';
 import 'package:lyte_studios_flutter_ui/theme/extensions/hex_color.dart';
 import 'package:readmore/readmore.dart';
 
-class CustomJobCard extends StatelessWidget {
+class CustomJobCard extends StatefulWidget {
+  final double height;
   final String description;
   final String userName;
   final String location;
@@ -15,9 +20,15 @@ class CustomJobCard extends StatelessWidget {
   final String suggestionPercentage;
   final VoidCallback onButtonPressed;
   final bool showBottomText;
+  final int descriptionPadding;
+  final bool isAICard;
+  final bool showLikeButton;
+
+  final double? rating;
 
   const CustomJobCard({
     super.key,
+    this.height = 272,
     required this.description,
     required this.userName,
     required this.location,
@@ -29,46 +40,141 @@ class CustomJobCard extends StatelessWidget {
     required this.suggestionPercentage,
     required this.onButtonPressed,
     required this.showBottomText,
+    this.descriptionPadding = 0,
+    this.isAICard = false,
+    this.showLikeButton = false,
+    this.rating,
   });
 
   @override
+  State<CustomJobCard> createState() => _CustomJobCardState();
+}
+
+class _CustomJobCardState extends State<CustomJobCard> {
+  bool isLiked = false;
+
+  void toggleLike() {
+    setState(() {
+      isLiked = !isLiked;
+    });
+  }
+
+  @override
   Widget build(BuildContext context) {
-    return Container(
-      height: 256,
-      width: MediaQuery.of(context).size.width * 0.8,
-      decoration: BoxDecoration(
-        color: HexColor.fromHex('#F6F6F6'),
-        borderRadius: BorderRadius.circular(20),
-      ),
-      child: Padding(
-        padding: const EdgeInsets.all(12),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
+    return GestureDetector(
+      onTap: () {
+        context.push(EmployProfileDisplayScreen.employerRoute);
+      },
+      child: Container(
+        constraints: const BoxConstraints(minHeight: 250),
+        decoration: BoxDecoration(
+          color: HexColor.fromHex('#F6F6F6'),
+          borderRadius: BorderRadius.circular(20),
+        ),
+        child: Stack(
           children: [
-            _buildProfileHeader(context),
-            const SizedBox(height: 10),
-            Expanded(child: _buildDescription(context)),
-            const SizedBox(height: 10),
-            _buildActionRow(context),
-            if (showBottomText == true) _buildBottomRow(context)
+            Padding(
+              padding: const EdgeInsets.all(12),
+              child: Column(
+                crossAxisAlignment: CrossAxisAlignment.start,
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  _buildProfileHeader(context),
+                  const SizedBox(height: 10),
+                  Flexible(
+                    fit: FlexFit.loose,
+                    child: Padding(
+                      padding: EdgeInsets.symmetric(
+                        horizontal: widget.descriptionPadding.toDouble(),
+                      ),
+                      child: _buildDescription(context),
+                    ),
+                  ),
+                  _buildActionRow(context),
+                ],
+              ),
+            ),
+            if (widget.showLikeButton)
+              Positioned(
+                top: 10,
+                right: 16,
+                child: Container(
+                  decoration: BoxDecoration(
+                    shape: BoxShape.circle,
+                    boxShadow: [
+                      BoxShadow(
+                        color: Colors.black.withOpacity(0.15),
+                        spreadRadius: 1,
+                        blurRadius: 4,
+                        offset: Offset(0, 2),
+                      ),
+                    ],
+                  ),
+                  child: CircleAvatar(
+                    backgroundColor: Colors.white,
+                    child: IconButton(
+                      color: const Color.fromRGBO(255, 255, 255, 1),
+                      icon: Icon(
+                        Icons.favorite,
+                        color: isLiked
+                            ? HexColor.fromHex('#FF3E68')
+                            : Colors.grey.shade300,
+                        size: 26,
+                      ),
+                      onPressed: toggleLike,
+                    ),
+                  ),
+                ),
+              ),
           ],
         ),
       ),
     );
   }
 
-  Row _buildProfileHeader(BuildContext context) {
+  Widget _buildProfileHeader(BuildContext context) {
     final theme = Theme.of(context);
     return Row(
+      crossAxisAlignment: CrossAxisAlignment.start,
       children: [
         _buildProfileImage(),
         const SizedBox(width: 10),
-        _buildUserInfo(theme),
+        Expanded(
+          child: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Text(
+                widget.userName,
+                style: TextStyle(
+                  fontFamily: 'Inter',
+                  fontSize: 18,
+                  fontWeight: FontWeight.w700,
+                  color: theme.colorScheme.onPrimaryContainer,
+                ),
+                maxLines: 1,
+                overflow: TextOverflow.ellipsis,
+              ),
+              const SizedBox(height: 4),
+              Row(
+                children: [
+                  _buildLocationInfo(),
+                  const SizedBox(width: 8),
+                  _buildAgeInfo(),
+                ],
+              ),
+              const SizedBox(height: 4),
+              if (widget.rating != null)
+                StarRating(
+                  rating: widget.rating!,
+                )
+            ],
+          ),
+        ),
       ],
     );
   }
 
-  Container _buildProfileImage() {
+  Widget _buildProfileImage() {
     return Container(
       width: 82,
       height: 82,
@@ -76,60 +182,39 @@ class CustomJobCard extends StatelessWidget {
         shape: BoxShape.circle,
         border: Border.all(color: Colors.white, width: 4),
         image: DecorationImage(
-          image: AssetImage(profileImagePath),
+          image: AssetImage(widget.profileImagePath),
           fit: BoxFit.cover,
         ),
       ),
     );
   }
 
-  Column _buildUserInfo(ThemeData theme) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Text(
-          userName,
-          style: TextStyle(
-            fontFamily: 'Inter',
-            fontSize: 18,
-            fontWeight: FontWeight.w700,
-            color: theme.colorScheme.onPrimaryContainer,
-          ),
-        ),
-        Row(
-          children: [
-            _buildLocationInfo(),
-            const SizedBox(width: 8),
-            _buildAgeInfo(),
-          ],
-        ),
-      ],
-    );
-  }
-
-  Row _buildLocationInfo() {
+  Widget _buildLocationInfo() {
     return Row(
       children: [
         Image.asset(
+          "assets/images/recruteren/location.png",
           height: 20,
           width: 20,
-          "assets/images/recruteren/location.png",
         ),
         const SizedBox(width: 4),
         Text(
-          location,
+          widget.location,
           style: TextStyle(
-            fontFamily: 'Inter',
+            fontFamily: 'Poppins',
             color: HexColor.fromHex('#666666'),
             fontSize: 15,
             fontWeight: FontWeight.w500,
+            letterSpacing: 0.0001,
           ),
+          maxLines: 1,
+          overflow: TextOverflow.ellipsis,
         ),
       ],
     );
   }
 
-  Row _buildAgeInfo() {
+  Widget _buildAgeInfo() {
     return Row(
       children: [
         Icon(
@@ -139,11 +224,12 @@ class CustomJobCard extends StatelessWidget {
         ),
         const SizedBox(width: 4),
         Text(
-          "$age jaar",
+          "${widget.age} jaar",
           style: TextStyle(
-            fontFamily: 'Inter',
+            fontFamily: 'Poppins',
             color: HexColor.fromHex('#666666'),
             fontSize: 15,
+            letterSpacing: 0.0001,
             fontWeight: FontWeight.w500,
           ),
         ),
@@ -151,20 +237,18 @@ class CustomJobCard extends StatelessWidget {
     );
   }
 
-  ReadMoreText _buildDescription(BuildContext context) {
+  Widget _buildDescription(BuildContext context) {
     final theme = Theme.of(context);
     return ReadMoreText(
-      description,
+      widget.description,
       trimLines: 3,
-      style: TextStyle(
-        fontFamily: 'Inter',
-        color: HexColor.fromHex('#4A4C53'),
+      style: TextStyles.bodyMedium.copyWith(
         fontSize: 16,
         fontWeight: FontWeight.w500,
       ),
       trimMode: TrimMode.Line,
-      trimCollapsedText: '\nleer meer',
-      trimExpandedText: '...Toon minder',
+      trimCollapsedText: '\nLeer meer',
+      trimExpandedText: ' Toon minder',
       lessStyle: TextStyle(
         fontFamily: 'Inter',
         color: theme.primaryColor,
@@ -180,24 +264,21 @@ class CustomJobCard extends StatelessWidget {
     );
   }
 
-  Row _buildActionRow(BuildContext context) {
+  Widget _buildActionRow(BuildContext context) {
     final theme = Theme.of(context);
     return Row(
       children: [
         Expanded(
           flex: 2,
-          child: CommonButton(
-            onButtonPressed: onButtonPressed,
-            buttonText: buttonText,
-            icon: buttonIcon,
-            backgroundColor: buttonColor,
+          child: ActionButton(
+            onButtonPressed: widget.onButtonPressed,
+            buttonText: widget.buttonText,
+            icon: widget.buttonIcon,
+            backgroundColor: widget.buttonColor,
           ),
         ),
         const SizedBox(width: 10),
-        Expanded(
-          flex: 1,
-          child: _buildSuggestionPercentage(theme),
-        ),
+        _buildSuggestionPercentage(theme),
       ],
     );
   }
@@ -208,33 +289,35 @@ class CustomJobCard extends StatelessWidget {
         Divider(
           color: HexColor.fromHex('#000000').withOpacity(0.1),
         ),
-        Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Text(
-              'Profiel bekijken',
-              style: TextStyle(
-                fontFamily: 'Inter',
-                color: HexColor.fromHex('#4A4C53'),
-                fontSize: 16,
-                fontWeight: FontWeight.w500,
+        ClearInkWell(
+          onTap: () {},
+          child: Row(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              Text(
+                'Profiel bekijken',
+                style: TextStyle(
+                  fontFamily: 'Inter',
+                  color: HexColor.fromHex('#4A4C53'),
+                  fontSize: 16,
+                  letterSpacing: -0.352,
+                  fontWeight: FontWeight.w500,
+                ),
               ),
-            ),
-            const SizedBox(
-              width: 5,
-            ),
-            Icon(
-              Icons.arrow_forward_ios,
-              size: 12,
-              color: HexColor.fromHex('#4A4C53'),
-            )
-          ],
+              const SizedBox(width: 5),
+              Icon(
+                Icons.arrow_forward_ios,
+                size: 12,
+                color: HexColor.fromHex('#4A4C53'),
+              ),
+            ],
+          ),
         ),
       ],
     );
   }
 
-  Container _buildSuggestionPercentage(ThemeData theme) {
+  Widget _buildSuggestionPercentage(ThemeData theme) {
     return Container(
       padding: const EdgeInsets.symmetric(horizontal: 12, vertical: 4),
       decoration: BoxDecoration(
@@ -242,15 +325,15 @@ class CustomJobCard extends StatelessWidget {
         border: Border.all(color: theme.primaryColor, width: 2),
       ),
       child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceAround,
         children: [
           Image.asset(
-            height: 18,
             "assets/images/recruteren/jobrAI_suggesties.png",
+            height: 18,
           ),
+          const SizedBox(width: 4),
           FittedBox(
             child: Text(
-              "$suggestionPercentage%",
+              " ${widget.suggestionPercentage}%",
               style: TextStyle(
                 fontFamily: 'Inter',
                 color: theme.primaryColor,
@@ -262,5 +345,67 @@ class CustomJobCard extends StatelessWidget {
         ],
       ),
     );
+  }
+}
+
+class StarRating extends StatelessWidget {
+  final double rating;
+  final int totalStars;
+  final double size;
+
+  const StarRating({
+    Key? key,
+    required this.rating,
+    this.totalStars = 5,
+    this.size = 24.0,
+  }) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      mainAxisSize: MainAxisSize.min,
+      children: [
+        Wrap(
+          spacing: -4,
+          children: List.generate(totalStars, (index) {
+            if (rating >= index + 1) {
+              return Icon(Icons.star_rounded,
+                  size: size, color: const Color(0xFFFFD400));
+            } else if (rating > index) {
+              final double fraction = rating - index;
+              return Stack(
+                children: [
+                  Icon(Icons.star_rounded,
+                      size: size, color: const Color(0xFFD9D9D9)),
+                  ClipRect(
+                    clipper: _StarClipper(fraction),
+                    child: Icon(Icons.star_rounded,
+                        size: size, color: const Color(0xFFFFD400)),
+                  ),
+                ],
+              );
+            } else {
+              return Icon(Icons.star_rounded,
+                  size: size, color: const Color(0xFFD9D9D9));
+            }
+          }),
+        ),
+      ],
+    );
+  }
+}
+
+class _StarClipper extends CustomClipper<Rect> {
+  final double fraction;
+  _StarClipper(this.fraction);
+
+  @override
+  Rect getClip(Size size) {
+    return Rect.fromLTWH(0, 0, size.width * fraction, size.height);
+  }
+
+  @override
+  bool shouldReclip(covariant _StarClipper oldClipper) {
+    return oldClipper.fraction != fraction;
   }
 }
